# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HomeHub household management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /family-members:
    get:
      operationId: getFamilyMembers
      tags: [family]
      summary: List all family members
      responses:
        "200":
          description: List of family members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
    post:
      operationId: createFamilyMember
      tags: [family]
      summary: Create a new family member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyMemberInput"
      responses:
        "201":
          description: Created family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
  /family-members/{id}:
    put:
      operationId: updateFamilyMember
      tags: [family]
      summary: Update a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFamilyMemberInput"
      responses:
        "200":
          description: Updated family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
        "404":
          description: Not found
    delete:
      operationId: deleteFamilyMember
      tags: [family]
      summary: Delete a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /chores:
    get:
      operationId: getChores
      tags: [chores]
      summary: List all chores
      parameters:
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of chores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chore"
    post:
      operationId: createChore
      tags: [chores]
      summary: Create a new chore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChoreInput"
      responses:
        "201":
          description: Created chore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
  /chores/{id}:
    put:
      operationId: updateChore
      tags: [chores]
      summary: Update a chore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChoreInput"
      responses:
        "200":
          description: Updated chore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
    delete:
      operationId: deleteChore
      tags: [chores]
      summary: Delete a chore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /chores/{id}/complete:
    post:
      operationId: completeChore
      tags: [chores]
      summary: Mark a chore as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteChoreInput"
      responses:
        "200":
          description: Updated chore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
...
components:
  schemas:
...
    FamilyMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [parent, child, pet]
        color:
          type: string
        photoUrl:
          type: string
          nullable: true
      required:
        - id
        - name
        - role
        - color
    CreateFamilyMemberInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [parent, child, pet]
        color:
          type: string
        photoUrl:
          type: string
          nullable: true
      required:
        - name
        - role
        - color
    UpdateFamilyMemberInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [parent, child, pet]
        color:
          type: string
        photoUrl:
          type: string
          nullable: true
...
    Chore:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        assigneeId:
          type: string
          nullable: true
        assigneeName:
          type: string
          nullable: true
        propertyId:
          type: string
        propertyName:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, biweekly, monthly, custom]
        dueDate:
          type: string
          format: date
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        completedBy:
          type: string
          nullable: true
        completionNote:
          type: string
          nullable: true
        isOverdue:
          type: boolean
        points:
          type: integer
      required:
        - id
        - title
        - propertyId
        - propertyName
        - frequency
        - isOverdue
        - points
    CreateChoreInput:
      type: object
      properties:
        title:
          type: string
        assigneeId:
          type: string
          nullable: true
        propertyId:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, biweekly, monthly, custom]
        dueDate:
          type: string
          format: date
          nullable: true
        points:
          type: integer
      required:
        - title
        - propertyId
        - frequency
    UpdateChoreInput:
      type: object
      properties:
        title:
          type: string
        assigneeId:
          type: string
          nullable: true
        propertyId:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, biweekly, monthly, custom]
        dueDate:
          type: string
          format: date
          nullable: true
        points:
          type: integer
    CompleteChoreInput:
      type: object
      properties:
        completedBy:
          type: string
        note:
          type: string
          nullable: true
      required:
        - completedBy
...